chattable command fixed

+++ app/Console/Commands/ChatTableFix.php
@@ -42,46 +42,57 @@ class ChatTableFix extends Command
      */
     public function handle()
     {
-        $msg_thread_table = DB::table('message_threads')->get();
+        if (Schema::hasTable('messages')) {
 
-        if ($msg_thread_table) {
-            foreach ($msg_thread_table as $thread) {
-                $this->line($thread->id);
-                $thread = Thread::create([
-                    'starred' => 1
-                ]);
-                $msg_thread_participants_table = DB::table('message_thread_participants')->where('thread_id', $thread->id)->get();
-                foreach ($msg_thread_participants_table as $participant) {
-                    Participant::create([
-                        'thread_id' => $thread->id,
-                        'user_id' => $participant->user_id,
-                        'last_read' => $participant->last_read
+            $msg_thread_table = DB::table('message_threads')->get();
+
+            if ($msg_thread_table) {
+                foreach ($msg_thread_table as $thread) {
+                    $this->line($thread->id);
+                    $thread = Thread::create([
+                        'starred' => 1
                     ]);
-                }
-                dump($msg_thread_participants_table);
-                $msg_messages_table = DB::table('messages')->where('thread_id', $thread->id)->get();
-                foreach ($msg_messages_table as $message) {
-                    Message::create([
-                        'thread_id' => $thread->id,
-                        'user_id' => $message->sender_id,
-                        'body' => $message->body,
-                        'created_at' => $message->created_at
-                    ]);
+                    $msg_thread_participants_table = DB::table('message_thread_participants')->where('thread_id', $thread->id)->get();
+                    foreach ($msg_thread_participants_table as $participant) {
+                        Participant::create([
+                            'thread_id' => $thread->id,
+                            'user_id' => $participant->user_id,
+                            'last_read' => $participant->last_read
+                        ]);
+                    }
+                    dump($msg_thread_participants_table);
+                    $msg_messages_table = DB::table('messages')->where('thread_id', $thread->id)->get();
+                    foreach ($msg_messages_table as $message) {
+                        Message::create([
+                            'thread_id' => $thread->id,
+                            'user_id' => $message->sender_id,
+                            'body' => $message->body,
+                            'created_at' => $message->created_at
+                        ]);
+                    }
                 }
             }
+
+            DB::statement('SET FOREIGN_KEY_CHECKS = 0');
+
+            if (Schema::hasTable('message_threads')) {
+                Schema::dropIfExists('message_threads');
+
+            }
+            if (Schema::hasTable('message_thread_participants')) {
+                Schema::dropIfExists('message_thread_participants');
+            }
+            if (Schema::hasTable('messages')) {
+                Schema::dropIfExists('messages');
+
+            }
+
+
+            unlink(database_path('/migrations/2016_07_27_052049_create_messages_table.php'));
+            unlink(database_path('/migrations/2016_07_31_215110_create_message_threads_table.php'));
+            unlink(database_path('/migrations/2016_07_31_215345_create_message_thread_participants.php'));
+            DB::statement('SET FOREIGN_KEY_CHECKS = 1');
         }
-
-        DB::statement('SET FOREIGN_KEY_CHECKS = 0');
-
-        Schema::dropIfExists('message_threads');
-        Schema::dropIfExists('message_thread_participants');
-        Schema::dropIfExists('messages');
-
-
-        unlink(database_path('/migrations/2016_07_27_052049_create_messages_table.php'));
-        unlink(database_path('/migrations/2016_07_31_215110_create_message_threads_table.php'));
-        unlink(database_path('/migrations/2016_07_31_215345_create_message_thread_participants.php'));
-        DB::statement('SET FOREIGN_KEY_CHECKS = 1');
 
         return $this->line("chat table imported");
     }
